+++ src/views/auth/signIn/index.jsx
@@ -108,9 +108,9 @@ function SignIn() {
   return (
     <>
       <PageHelmet
-        title="로그인 | Growth Analytics"
-        description="Growth Analytics 대시보드에 로그인하세요"
-        keywords="로그인, 그로스 애널리틱스, 대시보드"
+        title="로그인 | 제스트 애널리틱스"
+        description="제스트 애널리틱스 대시보드에 로그인하세요"
+        keywords="로그인, 제스트 애널리틱스, 대시보드"
       />
       <Flex
         w="100vw"